Fix ssprintf to not allocate temporary memory
The original code was written the way that it was because snprintf was not
standard and didn't work consistently across many platforms. But since C++11
std::snprintf is available and can be used to get the exact size of the
required buffer. Because of that, ssprintf can be greatly simplified.